Refactor RSDResourceTransformer to use a platform-specific resource loader
