+++ RSDCatalog/RSDCatalog/AppDelegate.swift
@@ -43,6 +43,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
         
+        resourceLoader = ResourceLoader()
+        
         // Set the default configuration for the catalog test app.
         RSDStudyConfiguration.shared.fullInstructionsFrequency = .daily
