+++ Research/ResearchUI/iOS/RSDAppDelegate.swift
@@ -67,6 +67,7 @@ open class RSDAppDelegate : UIResponder, RSDAppOrientationLock, RSDAlertPresente
         
         // Set up color palette and factory.
         RSDFactory.shared = instantiateFactory()
+        resourceLoader = ResourceLoader()
         let colorRules = instantiateColorRules()
         let fontRules = instantiateFontRules()
         if fontRules != nil || colorRules != nil {
